@@ -10,53 +10,51 @@
         "production": "mix --production"
     },
     "devDependencies": {
-        "@fullcalendar/bootstrap": "^5.10.0",
-        "@fullcalendar/daygrid": "^5.10.0",
-        "@fullcalendar/interaction": "^5.10.0",
-        "@fullcalendar/react": "^5.10.0",
-        "@fullcalendar/timegrid": "^5.10.0",
-        "@heroicons/react": "^1.0.5",
+        "@fullcalendar/bootstrap": "^6.0.0",
+        "@fullcalendar/daygrid": "^6.0.0",
+        "@fullcalendar/interaction": "^6.0.0",
+        "@fullcalendar/react": "^6.0.0",
+        "@fullcalendar/timegrid": "^6.0.0",
+        "@heroicons/react": "^1.0.6",
+        "@jridgewell/sourcemap-codec": "^1.4.14",
         "@svg-maps/world": "^1.0.1",
-        "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^11.2.7",
-        "@testing-library/user-event": "^12.8.3",
-        "@vitejs/plugin-react": "^1.3.2",
-        "apexcharts": "^3.29.0",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^14.4.3",
+        "@vitejs/plugin-react": "^2.2.0",
+        "apexcharts": "^3.36.3",
         "axios": "^1.1.2",
         "bootstrap": "^5.2.3",
-        "browser-sync": "^2.27.10",
-        "browser-sync-webpack-plugin": "^2.3.0",
-        "laravel-mix": "^5.0.9",
-        "leaflet": "^1.7.1",
+        "laravel-vite-plugin": "^0.7.1",
+        "leaflet": "^1.9.3",
         "lodash": "^4.17.19",
-        "moment-timezone": "^0.5.33",
+        "moment-timezone": "^0.5.40",
         "nouislider-react": "^3.4.1",
         "postcss": "^8.1.14",
-        "react": "^17.0.2",
-        "react-apexcharts": "^1.3.9",
-        "react-beautiful-dnd": "^13.1.0",
-        "react-bootstrap": "^2.0.1",
-        "react-bootstrap-table-next": "^4.0.3",
-        "react-bootstrap-table2-paginator": "^2.1.2",
-        "react-bootstrap-table2-toolkit": "^2.1.3",
-        "react-copy-to-clipboard": "^5.0.4",
-        "react-countup": "^6.0.0",
-        "react-datetime": "^3.1.1",
-        "react-device-detect": "^2.1.2",
-        "react-dom": "^17.0.2",
-        "react-dropzone": "^11.4.2",
-        "react-github-btn": "^1.2.1",
-        "react-leaflet": "^3.2.2",
-        "react-live": "^2.3.0",
+        "react": "^18.2.0",
+        "react-apexcharts": "^1.4.0",
+        "react-beautiful-dnd": "^13.1.1",
+        "react-bootstrap": "^2.7.0",
+        "react-copy-to-clipboard": "^5.1.0",
+        "react-countup": "^6.4.0",
+        "react-datetime": "^3.2.0",
+        "react-device-detect": "^2.2.2",
+        "react-dom": "^18.2.0",
+        "react-dropzone": "^14.2.3",
+        "react-github-btn": "^1.4.0",
+        "react-leaflet": "^4.2.0",
+        "react-live": "^3.1.1",
         "react-router-dom": "^5.3.4",
         "react-router-hash-link": "^2.4.3",
-        "react-scripts": "^3.4.3",
-        "react-select": "^5.1.0",
-        "react-transition-group": "^4.4.2",
-        "sass": "^1.56.1",
-        "simplebar-react": "^2.3.6",
-        "sweetalert2": "^11.1.9",
-        "sweetalert2-react-content": "^4.2.0",
-        "uuid": "^8.3.2"
+        "react-scripts": "^5.0.1",
+        "react-select": "^5.7.0",
+        "react-transition-group": "^4.4.5",
+        "rollup-plugin-terser": "^7.0.2",
+        "sass": "^1.56.2",
+        "simplebar-react": "^2.4.3",
+        "sweetalert2": "^11.6.15",
+        "sweetalert2-react-content": "^5.0.7",
+        "uuid": "^9.0.0",
+        "vite": "^3.2.5"
     }
 }
